+++ scratch.js
@@ -3,5 +3,7 @@ import spacetime from './src/index.js'
 
 // spacetime.extend(require('./plugins/holiday'))
 
-let s = spacetime.now('Etc/GMT+7')
-console.log(s.timezone()) //'Sábado'
+let s = spacetime('nov 11 2022', 'australia/adelaide')
+s = s.add(1, 'centuries')
+console.log(s.format())
+// console.log(s.season()) //'Sábado'